backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
SQLAlchemy==2.0.40
psycopg2-binary==2.9.10
Flask-WTF==1.2.2
Werkzeug==3.1.3

fpdf==1.7.2
PyPDF2==3.0.1
python-docx
pillow==11.2.1
matplotlib==3.10.1

sendgrid==6.11.0
twilio==9.5.2

python-dotenv
pandas
openpyxl

aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
courlan==1.3.2
cycler==0.12.1
dateparser==1.2.1
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.57.0
fpdf==1.7.2
frozenlist==1.6.0
greenlet==3.2.0
gunicorn==23.0.0
htmldate==1.9.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jusText==3.0.2
kiwisolver==1.4.8
lxml==5.4.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
matplotlib==3.10.1
multidict==6.4.3
numpy==2.2.5
packaging==25.0
pillow==11.2.1
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-http-client==3.3.7
pytz==2025.2
regex==2024.11.6
requests==2.32.3
sendgrid==6.11.0
six==1.17.0
SQLAlchemy==2.0.40
starkbank-ecdsa==2.2.0
tld==0.13
trafilatura==2.0.0
twilio==9.5.2
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.20.0
openpyxl
pandas
twilio
python-docx
python-dotenv
